完善redisson分布式锁功能减库存

+++ db-lock/src/main/java/com/appleyk/db/dao/mapper/CommodityStockMapper.java
@@ -17,7 +17,7 @@ public interface CommodityStockMapper extends BaseMapper<CommodityStock> {
 
     Integer reduce(@Param("code") String commodityCode);
 
-    Integer reduce2(@Param("code") String commodityCode,@Param("version") int version);
+    Integer reduce2(@Param("code") String commodityCode,@Param("version") long version);
 
     @Select("select commodity_code, commodity_name, inventory from t_commodity_stock where " +
             "commodity_code = #{commodityCode} for update")
